Populate homepage with JS not hard coded HTML
Removed the hard coded homepage design from index.html. Created
page-creation.js to contain modules that generate the page content for
each tab of the webpage. For now, this only contains a modle to generate
the content for the home page.

Modified the supporting webpack configuration files to include
functionality for asset management.