Bug fixes for creating the menu tab
Several fixes to resolve issues with the populateMenuPage module:
- Within the menu-items.csv, remove a whitespace preceeding the quoted
description field. This whitespace caused csv-loader to treat the cell
as if the content wasn't quoted, which was causing erroneous data for
this cell.
- Correct incorrect id attribute for the foodMenuTitle div. Now
correctly matches the format of the drinkMenuTitle id and the id
referenced in the stylesheet.
- Added unique id identifiers to the buttons in index.html. This will be
used to select the buttons in order to apply event listeners in a future
commit.
- Updated webpack config to dynamically type cell fields, and to
interpret the first row of data as a header row. This enables the
createMenuItems function within the module to function correctly.